wip: refactoring ios support

+++ src/apple/mac/streams/exception.rs
@@ -35,11 +35,11 @@ impl MinidumpWriter {
             .exception
             .as_ref()
             .map(|exc| {
-                let code = exc.code as u64;
+                let code = exc.code;
 
                 // `EXC_CRASH` exceptions wrap other exceptions, so we want to
                 // retrieve the _actual_ exception
-                let wrapped_exc = if exc.kind as u32 == et::EXC_CRASH {
+                let wrapped_exc = if exc.kind == et::EXC_CRASH {
                     recover_exc_crash_wrapped_exception(code)
                 } else {
                     None
@@ -62,7 +62,7 @@ impl MinidumpWriter {
                 // |[63:61] resource type | [60:58] flavor | [57:32] unused |
                 // +--------------------------------------------------------+
                 let exception_code =
-                    if exc.kind as u32 == et::EXC_RESOURCE || exc.kind as u32 == et::EXC_GUARD {
+                    if exc.kind == et::EXC_RESOURCE || exc.kind == et::EXC_GUARD {
                         (code >> 32) as u32
                     } else if let Some(wrapped) = wrapped_exc {
                         wrapped.code
@@ -169,6 +169,7 @@ struct WrappedException {
 /// Will return `None` if the specified code is wrapping an exception that
 /// should not be possible to be wrapped in an `EXC_CRASH`
 #[inline]
+#[allow(clippy::unnecessary_lazy_evaluations)]
 fn recover_exc_crash_wrapped_exception(code: u64) -> Option<WrappedException> {
     is_valid_exc_crash(code).then(|| WrappedException {
         kind: ((code >> 20) & 0xf) as u32,
